+++ examples/creator.js
@@ -1,5 +1,11 @@
-const P2PCommons = require('..') // liberate science constructor function
+const { join } = require('path')
+const {
+  promises: { writeFile }
+} = require('fs')
+const { encode } = require('dat-encoding')
+const once = require('events.once')
 const tempy = require('tempy')
+const P2PCommons = require('..') // liberate science constructor function
 const commons = new P2PCommons({
   baseDir: tempy.directory(),
   verbose: true,
@@ -10,10 +16,22 @@ process.once('SIGINT', () => commons.destroy())
 ;(async () => {
   await commons.ready()
   // create some content
-  const { rawJSON } = await commons.init({
+  const { rawJSON, driveWatch } = await commons.init({
     type: 'content',
     title: 'Cool 101',
     description: 'All the cool content you want to know',
+    main: 'file.txt'
+  })
+
+  await writeFile(
+    join(commons.baseDir, encode(rawJSON.url), 'file.txt'),
+    'hello'
+  )
+
+  await once(driveWatch, 'put-end')
+
+  await commons.set({
+    url: rawJSON.url,
     main: 'file.txt'
   })